requirements.txt
flask
stripe
python-telegram-bot
gunicorn